+++ src/Recyclarr.Cli/Console/Commands/ListQualitiesCommand.cs
@@ -9,7 +9,6 @@ using Spectre.Console.Cli;
 
 namespace Recyclarr.Cli.Console.Commands;
 
-#pragma warning disable CS8765
 [UsedImplicitly]
 [Description("List quality definitions in the guide for a particular service.")]
 internal class ListQualitiesCommand(